features/{shiftRequest, submittedShit, assignShift} has been moved to features/shift/{request, submit, assign}
