+++ src/features/shift/assign/index/get-by-shift-request-id/type.ts
@@ -1,0 +1,6 @@
+import type { AssignShift } from "@prisma/client";
+
+export interface GetAssigShiftResponse {
+	ok: true;
+	assignShift: AssignShift;
+}
