+++ src/features/shift/submit/route.ts
@@ -1,21 +1,21 @@
 import express from "express";
 import { attachStoreId } from "../../../middlewares/request/attachStoreId";
 import { attachUserId } from "../../../middlewares/request/attachUserId";
-// import { attachStoreIdFromHeader } from "../middlewares/request/attachStoreId";
-// import { attachUserIdFromCookie } from "../middlewares/request/cookie/attachUserIdFromCookie";
 import { validateshiftRequestId } from "../../../middlewares/validations/shiftRequestId.validate";
-import getAssignShiftController from "./index/get-by-shift-request-id/controller";
-import upsertAssignShfitController from "./index/post/controller";
+import getSubmittedShiftsSpesificController from "./all/controller";
+import upsertSubmittedShiftController from "./index/post/controller";
+import getSubmittedShiftUserController from "./me/controller";
 
 const router = express.Router();
 router.use(attachUserId);
 router.use(attachStoreId);
 
-router.post("/", upsertAssignShfitController);
+router.post("/", upsertSubmittedShiftController);
+router.get("/me", getSubmittedShiftUserController);
 router.get(
-	"/:shiftRequestId",
+	"/all/:shiftRequestId",
 	validateshiftRequestId,
-	getAssignShiftController,
+	getSubmittedShiftsSpesificController,
 );
 
 export default router;
